chat_room needs to be a POST parameter
For XHRs this is irrelevant anyway, but for "regular" requests the
parameter is confusing, because if there are multiple chat widgets on a
single page, all are supposed to be displayed; to distinguish which one
should be modified, we can use a POST parameter.  If there would be the
need for a query parameter, we would need a GET array.