Close #7: Improve UX and accessibility
It is important to presents users a clear document outline.
Unfortunately, the original HTML5 browser outline algorithm was never
properly implemented, and has been dropped from the specs.  Since we
cannot know where chat widgets will be placed, we cannot render
headings since these might break the outline, and we don't want users
to pass the desired heading levels to the plugin calls.  Thus we're
using a `<figure>` element for the whole chat widget, where we can use
a `<figcaption>` as accessible name[1].  We also use an `<ol>` instead
of a `<div>` for the chat history what is semantically meaningful.
Finally, we no longer hide the submit button of the form, and add a
label to the input.

[1] <https://developer.mozilla.org/en-US/docs/Glossary/Accessible_name>