Restrict max length of message to 160 characters
While 160 characters are a bit arbitrary, it is important to prevent
malicious users from posting overlong messages, so that the chat files
don't grow to large.  We handle this with HTML5 form validation, what
should be sufficient for regular users, and by just rejecting the post
if HTML5 form validation didn't catch that.

Obviously, this doesn't prevent users from posting huge amounts of
short messages, what causes the same problem as overlong messages, so
we will need to address this by other means.