Unify handling of XHR and regular requests
The most important change is that we now deliver the same HTML fragment
regardless of the type of request.  This allows us to handle errors the
same way, and we only need to distinguish post success (regular
requests need to redirect, but XHRs must not).

We also clean up a bit, and are more resilient regarding failed XHRs –
in this case it likely makes no sense to go on polling.