Increase default poll time
12 seconds can still cause quite some traffic, and is probably overkill
for the small chats; 60 seconds seems to be a more sensible default.

And yes, we should not poll but rather use notifications, but besides
that my time is limited, it might not be supported in all cases.  Okay,
we could at least use a more adaptable poll strategy, but again,
development is limited.

Anyhow, while we're at it, we apply phpcs to config and language files
as well.